+++ zproject/urls.py
@@ -169,6 +169,7 @@ from zerver.views.tutorial import set_tutorial_status
 from zerver.views.typing import send_notification_backend
 from zerver.views.unsubscribe import email_unsubscribe
 from zerver.views.upload import (
+    get_tusd_file_data,
     serve_file_backend,
     serve_file_download_backend,
     serve_file_unauthed_from_token,
@@ -368,6 +369,7 @@ v1_api_and_json_patterns = [
     rest_path("typing", POST=send_notification_backend),
     # user_uploads -> zerver.views.upload
     rest_path("user_uploads", POST=upload_file_backend),
+    rest_path("user_uploads/<file_id_str>", GET=(get_tusd_file_data, {"override_api_url_scheme"})),
     rest_path(
         "user_uploads/<realm_id_str>/<path:filename>",
         GET=(serve_file_url_backend, {"override_api_url_scheme"}),